Changed: mq components relocated to infrastructure layer
