+++ app/ingest/infrastructure/mq/stomp_interactor.py
@@ -7,8 +7,8 @@ from abc import ABC, abstractmethod
 
 import stomp
 
-from app.ingest.application.mq.mq_connection_params import MqConnectionParams
 from app.ingest.domain.mq.exceptions.mq_connection_exception import MqConnectionException
+from app.ingest.infrastructure.mq.mq_connection_params import MqConnectionParams
 
 
 class StompInteractor(ABC):
