+++ test/integration/ingest/infrastructure/mq/listeners/test_transfer_status_queue_listener.py
@@ -1,9 +1,9 @@
 import os
 from unittest.mock import patch
 
-from app.ingest.application.mq.listeners.transfer_status_queue_listener import TransferStatusQueueListener
-from app.ingest.application.mq.mq_connection_params import MqConnectionParams
-from test.integration.ingest.application.mq.listeners.stomp_listener_integration_test_base import \
+from app.ingest.infrastructure.mq.listeners.transfer_status_queue_listener import TransferStatusQueueListener
+from app.ingest.infrastructure.mq.mq_connection_params import MqConnectionParams
+from test.integration.ingest.infrastructure.mq.listeners.stomp_listener_integration_test_base import \
     StompListenerIntegrationTestBase
 
 
@@ -23,7 +23,7 @@ class TestTransferStatusQueueListener(StompListenerIntegrationTestBase):
     def tearDown(self) -> None:
         self.sut.disconnect()
 
-    @patch("app.ingest.application.mq.listeners.transfer_status_queue_listener.TransferStatusQueueListener"
+    @patch("app.ingest.infrastructure.mq.listeners.transfer_status_queue_listener.TransferStatusQueueListener"
            "._handle_received_message")
     def test_on_message_happy_path(self, on_message_mock) -> None:
         self._send_test_message(self.TEST_MESSAGE)
