+++ test/integration/ingest/infrastructure/mq/stomp_integration_test_base.py
@@ -3,7 +3,7 @@ from abc import abstractmethod, ABC
 
 import stomp
 
-from app.ingest.application.mq.mq_connection_params import MqConnectionParams
+from app.ingest.infrastructure.mq.mq_connection_params import MqConnectionParams
 from test.integration.integration_test_base import IntegrationTestBase
